+++ superduperdb/base/code.py
@@ -6,10 +6,11 @@ from superduperdb.base.serializable import Serializable
 
 template = \
 """from superduperdb import code
+
 @code
 {definition}"""
 
-default = template.format(definition='def my_code(*args, **kwargs):\n    ...\n    return\n')
+default = template.format(definition='def my_code(x):\n    return x\n')
 
 
 @dc.dataclass