+++ superduperdb/ext/numpy/encoder.py
@@ -25,7 +25,8 @@ class DecodeArray:
         return numpy.frombuffer(bytes, dtype=self.dtype).reshape(self.shape)
 
 
-def array(dtype: str, shape: t.Sequence, bytes_encoding: t.Optional[str] = None):
+def array(dtype: str, shape: t.Sequence, bytes_encoding: t.Optional[str] = None,
+          encodable: str = 'encodable'):
     """
     Create an encoder of numpy arrays.
 
@@ -38,4 +39,5 @@ def array(dtype: str, shape: t.Sequence, bytes_encoding: t.Optional[str] = None)
         decoder=DecodeArray(dtype, shape),
         shape=shape,
         bytes_encoding=bytes_encoding,
+        encodable=encodable,
     )