+++ superduperdb/ext/sentence_transformers/model.py
@@ -1,6 +1,8 @@
 import dataclasses as dc
 import typing as t
 
+from superduperdb.base.code import Code
+from superduperdb.components.component import ensure_initialized
 from superduperdb.misc.annotations import requires_packages
 
 from sentence_transformers import SentenceTransformer as _SentenceTransformer
@@ -8,6 +10,11 @@ from sentence_transformers import SentenceTransformer as _SentenceTransformer
 from superduperdb.backends.query_dataset import QueryDataset
 from superduperdb.components.datatype import DataType, dill_serializer
 from superduperdb.components.model import Model, Signature, _DeviceManaged
+
+
+DEFAULT_PREDICT_KWARGS = {
+    'show_progress_bar': True,
+}
 
 
 @dc.dataclass(kw_only=True)
@@ -19,9 +26,24 @@ class SentenceTransformer(Model, _DeviceManaged):
     object: t.Optional[_SentenceTransformer] = None
     model: t.Optional[str] = None
     device: str = 'cpu'
-    preprocess: t.Optional[t.Callable] = None
-    postprocess: t.Optional[t.Callable] = None
+    preprocess: t.Union[None, t.Callable, Code] = None
+    postprocess: t.Union[None, t.Callable, Code] = None
     signature: Signature = 'singleton'
+
+    ui_schema: t.ClassVar[t.List[t.Dict]] = [
+        {'name': 'model', 'type': 'str', 'default': 'all-MiniLM-L6-v2'},
+        {'name': 'device', 'type': 'str', 'default': 'cpu', 'choices': ['cpu', 'cuda']},
+        {'name': 'predict_kwargs', 'type': 'json', 'default': DEFAULT_PREDICT_KWARGS},
+        {'name': 'postprocess', 'type': 'code', 'default': Code.default},
+    ]
+
+    @classmethod
+    def handle_integration(cls, kwargs):
+        if isinstance(kwargs.get('preprocess'), str):
+            kwargs['preprocess'] = Code(kwargs['preprocess'])
+        if isinstance(kwargs.get('postprocess'), str):
+            kwargs['postprocess'] = Code(kwargs['postprocess'])
+        return kwargs
 
     def __post_init__(self, artifacts):
         super().__post_init__(artifacts)
@@ -30,7 +52,7 @@ class SentenceTransformer(Model, _DeviceManaged):
             self.model = self.identifier
 
         if self.object is None:
-            self.object = _SentenceTransformer(self.identifier, device=self.device)
+            self.object = _SentenceTransformer(self.model, device=self.device)
 
         self.to(self.device)
 
@@ -38,6 +60,7 @@ class SentenceTransformer(Model, _DeviceManaged):
         self.object = self.object.to(device)
         self.object._target_device = device
 
+    @ensure_initialized
     def predict_one(self, X, *args, **kwargs):
         if self.preprocess is not None:
             X = self.preprocess(X)
@@ -48,6 +71,7 @@ class SentenceTransformer(Model, _DeviceManaged):
             result = self.postprocess(result)
         return result
 
+    @ensure_initialized
     def predict(self, dataset: t.Union[t.List, QueryDataset]) -> t.List:
         if self.preprocess is not None:
             dataset = list(map(self.preprocess, dataset))  # type: ignore[arg-type]