+++ superduperdb/misc/annotations.py
@@ -178,8 +178,10 @@ def _get_indent(docstring: str) -> int:
     return len(non_empty_lines[1]) - len(non_empty_lines[1].lstrip())
 
 
-def ui(*schema: t.Sequence[t.Dict]):
+def ui(*schema: t.Sequence[t.Dict], handle_integration: t.Callable = lambda x: x):
     def decorated(f):
         f.get_ui_schema = lambda: schema
+        f.build = lambda r: f(**r)
+        f.handle_integration = handle_integration
         return f
     return decorated