+++ test/conftest.py
@@ -114,9 +114,9 @@ def test_db(request) -> Iterator[Datalayer]:
             db.databackend.conn.drop_database(db_name)
             db.databackend.conn.drop_database(f'_filesystem:{db_name}')
         except Exception as e:
-            logging.warning(f"Error dropping databases: {e}")
-            for c in db.databackend.db.list_collections():
-                getattr(db.databackend.db, c).delete_many({})
+            logging.info(f"Error dropping databases: {e}")
+            for c in db.databackend.db.list_collection_names():
+                db.databackend.db.drop_collection(c)
 
 
 @pytest.fixture