+++ test/unittest/base/test_document.py
@@ -13,7 +13,7 @@ try:
 except ImportError:
     torch = None
 
-from superduperdb.base.document import Document
+from superduperdb.base.document import Document, _build_leaves
 
 
 @pytest.fixture
@@ -59,3 +59,42 @@ def test_underscore_artifacts(db):
 
     import PIL.PngImagePlugin
     assert isinstance(r['x'], PIL.PngImagePlugin.PngImageFile)
+
+
+def my_function(x):
+    return x + 2
+
+@pytest.mark.parametrize(
+    "db", [DBConfig.mongodb_empty], indirect=True
+)
+def test_build_leaves(db):
+    # TODO - this is how we encode data going forward
+
+    raw = db.datatypes['dill'].encoder(my_function)
+    file_id = db.artifact_store.save_artifact({'bytes': raw, 'datatype': 'dill'})['file_id']
+
+    leaf_records = [
+        {
+            'leaf_type': 'artifact', 
+            'cls': 'Artifact', 
+            'module': 'superduperdb.components.datatype', 
+            'dict': {
+                'file_id': file_id,
+                'datatype': 'dill',
+            }
+        },
+        {
+            'leaf_type': 'component', 
+            'cls': 'ObjectModel', 
+            'module': 'superduperdb.components.model', 
+            'dict': {'identifier': 'test', 'object': f'_artifact/{file_id}'}
+        },
+        {
+            'leaf_type': 'component', 
+            'cls': 'Stack', 
+            'module': 'superduperdb.components.stack', 
+            'dict': {'identifier': 'test_stack', 'components': ['_component/model/test_']}
+        },
+    ]
+    out = _build_leaves(leaf_records)
+    print(out)