+++ test/unittest/component/test_stack.py
@@ -1,0 +1,28 @@
+from superduperdb import superduper
+from superduperdb.base.document import _build_leaves
+from superduperdb.components.stack import Stack
+
+db = superduper('mongomock://localhost:27017/test_db')
+
+defn = {
+    "identifier": "mystack",
+    "_leaves": [
+        {
+            "leaf_type": "component",
+            "cls": "image_type",
+            "module": "superduperdb.ext.pillow.encoder",
+            "dict": {
+                "identifier": "my_image",
+                "media_type": "image/png"
+            }
+        }
+    ]
+}
+
+out, exit = _build_leaves(defn["_leaves"], db)
+
+out2 = Stack.from_list(
+    content=defn["_leaves"],
+    identifier=defn["identifier"],
+    db=db,
+)